sudoku-grid-board correction
